backend/pyproject.toml
[tool.poetry]
name = "llmkg-backend"
version = "0.1.0"
description = "llmkg backend"
authors = ["Harry Sadler <harrysadlermusic@gmail.com>"]
package-mode = false
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "*"
uvicorn = {extras = ["standard"], version = "^0.22.0"}
python-json-logger = "*"
asyncpg = "*"
alembic = "*"
psycopg2-binary = "*"
prometheus-fastapi-instrumentator = "*"
openai = "*"
types-requests = "*"
pydantic-settings = "*"

[tool.poetry.group.dev.dependencies]
black = "*"
mypy = "*"
flake8 = "*"
flake8-isort = "*"
pytest = "*"
pytest-mock = "*"
pytest-asyncio = "*"

[tool.black]
line-length = 100

[tool.isort]
profile="black"
line_length = 100

[tool.mypy]
strict = true
ignore_missing_imports = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
